+++ packages/cache/RELEASES.md
@@ -11,4 +11,4 @@
 - Fix to await async function getCompressionMethod
 
 ### 0.3.0
-- Downloads Azure-hosted caches using AzCopy when available
+- Downloads Azure-hosted caches using the Azure SDK for speed and reliability